Update Kotter to latest stable
This includes a fix to running multiple separate sessions
(before, earlier sessions wouldn't get cleaned up properly
and would affect later sessions).

This will be useful for us in a followup commit as we plan
to kick off a separate session for finding the Kobweb folder
if not immediately found.