If the curr dir is not a Kobweb application, we now search a little before giving up
More specifically, if the user is one or two folder too high, we'll search their
subdirectories to see if we can find a .kobweb folder in there instead.

This should help new users who get confused when the run `kobweb run` in a project's
root instead of the correct site subfolder.

We limit the search to depth = 2 for now to prevent going crazily deep if the user
runs a kobweb command in a totally incorrect location by accident.

When a .kobweb folder is not found, there are three possible cases:

1) We can't find a Kobweb application project even AFTER searching.
Outcome: Show the same error message as before.

2) We find exactly one Kobweb application under the current directory.
Outcome: Ask user a yes/no question.

3) We find multiple Kobweb projects.
Outcome: Ask the user to choose from a list of options.

Fixes #6