Improve CLI messaging around relative paths when Kobweb project can't be found
The new behavior shows a path relative to the current directory the user is in.
This results in a more intuitive message in the following case...

```
cd ~/Code/kobweb
kobweb run -p playground

Kobweb found a project instead in playground/site. Use it? [Yes]

Running: kobweb run -p playground/site
```

Before, we were using the path relative to the -p argument, which would have
resulted in:

```
Running: kobweb run -p site
```
which is wrong given where the user was calling kobweb from.