Add support for maximum version in template metadata
This can allow us to do something like this:

```
templates/app/
   1.2.2/ # max version set to 1.2.2, name set to "app"
   1.2.3/ # min version set to 1.2.3, name set to "app"
```
at which point, if we add a new feature to CLI 1.2.3, we can
fork 1.2.2 into a new path and use it there. That way, old
users will see the old app and new users will see the new
app.