Add `not` to freemarker methods
This can be useful for negating conditions in the kobweb templates,
e.g. converting a condition like `{useWorker}` to `{not(useWorker)}`
(since using ! doesn't work easily since `useWorker` in this case is
actually a string)

There may be ways to use raw freemarker to accomplish the same thing,
but I struggled for a while and could not get it not to spew out
errors. This approach seems nice enough anyway.