Protect against parsing exceptions in templates
This is not something we're actively experiencing right now, but
it's something we should have protected against from the beginning.
For example, maybe with Kobweb 1.1.0, we need to change the template
format in some drastic way that would cause all old templates to
crash. Ideally that wouldn't happen, but now we would just ignore
those old templates rather that totally crashing the CLI.