Change "--layout kobweb" to "--layout fullstack"
The word "kobweb" is legacy and results in a confusing exception
given that docs always talk about "static vs fullstack" layouts.

So let's remove the confusion! And use terms more consistently.

We can do this since Kobweb 0.17.1, which introduced the first
change. Of course, we need to be careful, because this could
break if people use this new CLI to target a project using an
older project. Therefore, we should probably wait a little
whlie before releasing this.