Simplify CLI code around static layout + dev environment
The Kobweb server design used to say that running a
static layout server in dev mode was not valid. However,
that is changing in 0.17.2. So we remove the checks here
that prevent such a combo from going in.

Note that even if a user runs a newer CLI with an older
project, that it will still work fine, because the server
will handle it anyway. It's possible we really never should
have even shown the warning in the first place.