Update confusing and incorrect error message about deleted conf.yaml
It turns out that if a conf.yaml file cannot be parsed, we were interpreting that
as the file not existing, which caused us to deliver a confusing error message.

Now, we distinguish between the two cases and issue an appropriate error based
on whether the file was deleted or just not parsable.

Bug #14