Visual tweaks to formatting
From this:

```
Press [Q] anytime to stop the server. Press [H] for help.

[H]elp   Toggle this help view
[Q]uit   Shutdown the server
[P]ause  Toggle live reloading [ON]
```

to this:

```
Press Q anytime to stop the server. Press H for help.

Commands:

  h  help   Toggle this help view
  q  quit   Shutdown the server
  p  pause  Toggle live reloading [ON]
```

Removing the "[]" feels cleaner and doesn't visually conflict
with the fact that we use square brackets for selecting options,
as in "Do the thing? [yes]  no "